--- tests/test_voyager/test_voyager_target_pmburn2.py
@@ -1,28 +1,0 @@
-"""
-Describe purpose of this script here
-
-Created: 1/25/25
-"""
-import pytest
-from spiceypy import kclear
-
-from vehicle.voyager import init_spice
-from vehicle.voyager_depart_target import PMTargeter
-
-
-@pytest.fixture
-def kernels():
-    init_spice()
-    yield None
-    kclear()
-
-
-@pytest.mark.parametrize(
-    "vgr_id",
-    [1,2]
-)
-def test_target_pm(tmp_path,kernels,vgr_id):
-    pm_targeter=PMTargeter(vgr_id=vgr_id,out_path=tmp_path)
-    pm_targeter.target(optimize=False)
-
-
